+++ CHANGELOG.md
@@ -5,8 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]
 ----------------------
+##### Changed
+- Removed animation when removing a repeat.
+
 ##### Fixed
 - When a `setvalue` element has no `value` attribute and no textContent, it does not work for resetting values(enketo-transformer fix).
+- When nested repeats using jr:count have values that empty or 0, a nested repeat can never be created (exception)
 
 [5.16.6] - 2020-08-03
 ----------------------